+++ lib/modules/event/campaign_detail_page.dart
@@ -1,6 +1,7 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:chaldea/components/components.dart';
 import 'package:chaldea/modules/shared/item_related_builder.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 import 'event_base_page.dart';
 
@@ -110,10 +111,7 @@ class _CampaignDetailPageState extends State<CampaignDetailPage>
               PopupMenuItem(
                 child: Text(S.current.jump_to('Mooncell')),
                 onTap: () {
-                  jumpToExternalLinkAlert(
-                    url: WikiUtil.mcFullLink(widget.event.indexKey),
-                    name: 'Mooncell',
-                  );
+                  launch(WikiUtil.mcFullLink(widget.event.indexKey));
                 },
               )
             ],