+++ lib/modules/statistics/game_statistics_page.dart
@@ -1,6 +1,7 @@
 import 'package:chaldea/components/components.dart';
 import 'package:chaldea/modules/item/item_detail_page.dart';
 import 'package:chaldea/modules/shared/item_related_builder.dart';
+import 'package:chaldea/modules/shared/list_page_share.dart';
 
 import 'statistics_servant_tab.dart';
 
@@ -12,12 +13,11 @@ class GameStatisticsPage extends StatefulWidget {
 class _GameStatisticsPageState extends State<GameStatisticsPage>
     with SingleTickerProviderStateMixin {
   late TabController _tabController;
-  Map<String, int>? allItemCost;
 
   @override
   void initState() {
     super.initState();
-    _tabController = TabController(length: 2, vsync: this);
+    _tabController = TabController(length: 3, vsync: this);
   }
 
   @override
@@ -32,47 +32,94 @@ class _GameStatisticsPageState extends State<GameStatisticsPage>
       appBar: AppBar(
         leading: BackButton(),
         title: Text(S.of(context).statistics_title),
-        actions: [],
-        bottom: TabBar(controller: _tabController, tabs: [
-          Tab(text: S.of(context).item),
-          Tab(text: S.of(context).servant)
-        ]),
+        actions: [
+          buildSwitchPlanButton(
+            context: context,
+            onChange: (index) {
+              db.curUser.curSvtPlanNo = index;
+              db.itemStat.update();
+              setState(() {});
+            },
+          )
+        ],
+        bottom: TabBar(
+          controller: _tabController,
+          isScrollable: true,
+          tabs: [
+            Tab(
+                text: LocalizedText.of(
+                    chs: '素材需求', jpn: 'アイテム需要', eng: 'Item Demands')),
+            Tab(
+                text: LocalizedText.of(
+                    chs: '已消耗素材', jpn: 'アイテム消費済', eng: 'Item Consumed')),
+            Tab(text: S.of(context).servant)
+          ],
+        ),
       ),
       body: TabBarView(
         controller: _tabController,
         // pie chart relate
         physics: AppInfo.isMobile ? NeverScrollableScrollPhysics() : null,
         children: [
-          // [PrimaryScrollerController]
-          KeepAliveBuilder(builder: (context) => _buildItemTab()),
+          KeepAliveBuilder(builder: (context) => StatItemDemandsTab()),
+          KeepAliveBuilder(builder: (context) => StatItemConsumedTab()),
           KeepAliveBuilder(builder: (context) => StatisticServantTab())
         ],
       ),
     );
   }
+}
 
-  bool includeCurItems = false;
+class StatItemConsumedTab extends StatefulWidget {
+  const StatItemConsumedTab({Key? key}) : super(key: key);
 
-  Widget _buildItemTab() {
+  @override
+  _StatItemConsumedTabState createState() => _StatItemConsumedTabState();
+}
+
+class _StatItemConsumedTabState extends State<StatItemConsumedTab> {
+  late ScrollController _scrollController;
+  Map<String, int> shownItems = {};
+  bool includeOwnedItems = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _scrollController = ScrollController();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _scrollController.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
     calculateItem();
-    final shownItems =
-        sumDict([allItemCost, if (includeCurItems) db.curUser.items]);
-    shownItems.removeWhere((key, value) {
-      int group = (db.gameData.items[key]?.id ?? 0) ~/ 100;
-      return key != Items.qp && (!(group >= 10 && group < 40) || value <= 0);
-    });
     return ListView(
-      padding: EdgeInsets.symmetric(vertical: 12),
+      controller: _scrollController,
+      padding: EdgeInsets.symmetric(vertical: 8),
       children: [
-        CheckboxListTile(
-          value: includeCurItems,
-          onChanged: (v) => setState(() {
-            if (v != null) includeCurItems = v;
-          }),
-          controlAffinity: ListTileControlAffinity.leading,
-          title: Text(S.of(context).statistics_include_checkbox),
+        Wrap(
+          spacing: 6,
+          runSpacing: 6,
+          alignment: WrapAlignment.center,
+          children: [
+            CheckboxWithLabel(
+              value: includeOwnedItems,
+              label: Text(LocalizedText.of(
+                  chs: '包含库存', jpn: '在庫を含める', eng: 'Include Owned')),
+              onChanged: (v) {
+                setState(() {
+                  if (v != null) includeOwnedItems = v;
+                });
+              },
+            ),
+          ],
         ),
         CustomTile(
+          color: Theme.of(context).cardColor,
           leading: db.getIconImage(Items.qp, height: 56),
           title: Text(formatNumber(shownItems[Items.qp] ?? 0)),
           onTap: () =>
@@ -91,8 +138,7 @@ class _GameStatisticsPageState extends State<GameStatisticsPage>
   }
 
   void calculateItem() {
-    if (allItemCost != null) return;
-    allItemCost = {};
+    shownItems.clear();
     final emptyPlan = ServantPlan(favorite: true);
     db.curUser.servants.forEach((no, svtStat) {
       if (!svtStat.favorite) return;
@@ -102,9 +148,119 @@ class _GameStatisticsPageState extends State<GameStatisticsPage>
       }
       final svt = db.gameData.servantsWithUser[no]!;
       sumDict(
-        [allItemCost, svt.getAllCost(cur: emptyPlan, target: svtStat.curVal)],
+        [shownItems, svt.getAllCost(cur: emptyPlan, target: svtStat.curVal)],
         inPlace: true,
       );
+    });
+    sumDict([shownItems, if (includeOwnedItems) db.curUser.items],
+        inPlace: true);
+    shownItems.removeWhere((key, value) {
+      int group = (db.gameData.items[key]?.id ?? 0) ~/ 100;
+      return key != Items.qp && (!(group >= 10 && group < 40) || value <= 0);
+    });
+  }
+}
+
+class StatItemDemandsTab extends StatefulWidget {
+  const StatItemDemandsTab({Key? key}) : super(key: key);
+
+  @override
+  _StatItemDemandsTabState createState() => _StatItemDemandsTabState();
+}
+
+class _StatItemDemandsTabState extends State<StatItemDemandsTab> {
+  late ScrollController _scrollController;
+  Map<String, int> shownItems = {};
+  bool subtractOwnedItems = false;
+  bool subtractEventItems = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _scrollController = ScrollController();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _scrollController.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    calculateItem();
+    return ListView(
+      controller: _scrollController,
+      padding: EdgeInsets.symmetric(vertical: 8),
+      children: [
+        Wrap(
+          spacing: 6,
+          runSpacing: 6,
+          alignment: WrapAlignment.center,
+          children: [
+            CheckboxWithLabel(
+              value: subtractOwnedItems,
+              label: Text(LocalizedText.of(
+                  chs: '减去库存', jpn: '在庫を差し引く', eng: 'Subtract Owned')),
+              onChanged: (v) {
+                setState(() {
+                  subtractOwnedItems = v ?? subtractOwnedItems;
+                });
+              },
+            ),
+            CheckboxWithLabel(
+              value: subtractEventItems,
+              label: Text(LocalizedText.of(
+                  chs: '减去活动所得', jpn: '活動収入を差し引く', eng: 'Subtract Event')),
+              onChanged: (v) {
+                setState(() {
+                  subtractEventItems = v ?? subtractEventItems;
+                });
+              },
+            ),
+          ],
+        ),
+        CustomTile(
+          color: Theme.of(context).cardColor,
+          leading: db.getIconImage(Items.qp, height: 56),
+          title: Text(formatNumber(shownItems[Items.qp] ?? 0)),
+          onTap: () =>
+              SplitRoute.push(context, ItemDetailPage(itemKey: Items.qp)),
+        ),
+        buildClassifiedItemList(
+          context: context,
+          data: shownItems..remove(Items.qp),
+          divideRarity: true,
+          divideClassItem: false,
+          compactNum: false,
+          minCrossCount: 7,
+        )
+      ],
+    );
+  }
+
+  void calculateItem() {
+    shownItems.clear();
+    db.curUser.servants.forEach((no, svtStat) {
+      if (!svtStat.favorite) return;
+      final svt = db.gameData.servantsWithUser[no];
+      if (svt == null) {
+        print('No $no: ${db.gameData.servantsWithUser.length}');
+        return;
+      }
+      sumDict([
+        shownItems,
+        svt.getAllCost(cur: svtStat.curVal, target: db.curUser.svtPlanOf(no))
+      ], inPlace: true);
+    });
+    sumDict([
+      shownItems,
+      if (subtractOwnedItems) multiplyDict(db.curUser.items, -1),
+      if (subtractEventItems) multiplyDict(db.itemStat.eventItems, -1),
+    ], inPlace: true);
+    shownItems.removeWhere((key, value) {
+      int group = (db.gameData.items[key]?.id ?? 0) ~/ 100;
+      return key != Items.qp && (!(group >= 10 && group < 40) || value <= 0);
     });
   }
 }