+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - sort by bond or item efficiency in free quest calculator
 - fix popup menu not show dialog/new route
 - more profile support for Altoria Caster 
+- servant equipped cmd code planning
+- add item demands/consumed overview in Statistics
 
 ## 1.5.1
 - Import servant/item data from http://fgosimulator.webcrow.jp/Material/